+++ module/core/variadic_from/tests/inc/auto_std_named_derive.rs
@@ -1,0 +1,17 @@
+#[ allow( unused_imports ) ]
+use super::*;
+
+#[ allow( unused_imports ) ]
+use the_module::exposed::*;
+
+#[ derive( Debug, PartialEq, Default, VariadicFrom ) ]
+struct StructNamedFields
+{
+  a : i32,
+  b : i32,
+}
+
+// Standard From and Into auto derive From_1 and To_1.
+
+include!( "./only_test/variadic_from2_named.rs" );
+include!( "./only_test/variadic_std_from2_named.rs" );
