+++ module/core/variadic_from/tests/inc/mod.rs
@@ -1,7 +1,11 @@
 use super::*;
 
 #[ cfg( all( feature = "type_variadic_from" ) ) ]
-mod variadic_from_manual_test;
+mod from4_named_manual;
+#[ cfg( all( feature = "type_variadic_from" ) ) ]
+mod from4_tuple_manual;
+
+// mod variadic_from_manual_test;
 
 #[ cfg( all( feature = "type_variadic_from" ) ) ]
 mod variadic_from_manual_beyond_test;
@@ -10,5 +14,9 @@ mod variadic_from_manual_beyond_test;
 // mod variadic_from_derive_test;
 // xxx : fix
 
+// xxx
+// #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
+// mod auto_std_named_derive;
+
 #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
-mod variadic_from2_derive;
+mod auto_std_named_manual;
