+++ module/core/variadic_from/tests/inc/only_test/variadic_from2_named.rs
@@ -1,3 +1,5 @@
+#[ test ]
+fn variadic_from2_named()
 {
 
   //
